Remove old pricing code
