+++ Sources/PointFree/SubscribeConfirmation.swift
@@ -1,3 +1,4 @@
+import Either
 import HttpPipeline
 import Foundation
 import Models
@@ -66,4 +67,55 @@ private func fetchCoupon(_ couponId: Stripe.Coupon.Id?) -> IO<Stripe.Coupon?> {
   return Current.stripe.fetchCoupon(couponId)
     .run
     .map(^\.right)
+}
+
+func redirectActiveSubscribers<A>(
+  user: @escaping (A) -> User?
+  )
+  -> (@escaping Middleware<StatusLineOpen, ResponseEnded, A, Data>)
+  -> Middleware<StatusLineOpen, ResponseEnded, A, Data> {
+
+    return { middleware in
+      return { conn in
+        let user = user(conn.data)
+
+        let userSubscription = (user?.subscriptionId)
+          .map(
+            Current.database.fetchSubscriptionById
+              >>> mapExcept(requireSome)
+          )
+          ?? throwE(unit)
+
+        let ownerSubscription = (user?.id)
+          .map(
+            Current.database.fetchSubscriptionByOwnerId
+              >>> mapExcept(requireSome)
+          )
+          ?? throwE(unit)
+
+        let race = (userSubscription.run.parallel <|> ownerSubscription.run.parallel).sequential
+
+        return EitherIO(run: race)
+          .flatMap {
+            $0.stripeSubscriptionStatus == .canceled
+              ? throwE(unit as Error)
+              : pure($0)
+          }
+          .run
+          .flatMap(
+            either(
+              const(
+                middleware(conn)
+              ),
+              const(
+                conn
+                  |> redirect(
+                    to: .account(.index),
+                    headersMiddleware: flash(.warning, "You already have an active subscription.")
+                )
+              )
+            )
+        )
+      }
+    }
 }
